+++ sample/src/com/manuelpeinado/imagelayout/demo/HomeActivity.java
@@ -34,7 +34,8 @@ public class HomeActivity extends ListActivity {
             new ActivityInfo(BasicUsageActivity.class, R.string.activity_title_basic_usage),
             new ActivityInfo(AddChildrenProgrammaticallyActivity.class, R.string.activity_title_add_children_programmatically),
             new ActivityInfo(FitAttributeActivity.class, R.string.activity_title_fit_attribute),
-            new ActivityInfo(HorizontalScrollViewActivity.class, R.string.activity_title_horizontal_scroll_view));
+            new ActivityInfo(HorizontalScrollViewActivity.class, R.string.activity_title_horizontal_scroll_view),
+            new ActivityInfo(ChangeImageDynamicallyActivity.class, R.string.activity_title_change_image_dynamically));
     
     @Override
     protected void onCreate(Bundle savedInstanceState) {